+++ io_scene_xray/object_/fmt_object.py
